Simplified log extensions
